introduce image compression
This commit introduces image compression, done via a library found on github. This compression is made on each picture imported from standard picker, and when done the app is automatically closed. For now, parameters are hardcoded but we need to provide the possibility to change them dinamically from the app settings.