introduce multi-selection list
this commit introduces the multi-selection feature, which enables the deletion of elements in the list of retrieved images. We're making use of another library, which is not supporting well all the functionalities, so some workaround was inserted. Moreover, we moved some functions to FileUtils class in order to split up and clean the code. Successive commits will clean the MainActivity.kt code more.